backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
cartesia==2.0.0
pydub==0.25.1
aiohttp==3.10.5
httpx==0.27.2
httpx-sse==0.4.0
iterators==0.3.0
pydantic==2.9.2
pydantic-core==2.23.4
typing_extensions==4.12.2
websockets==13.1